Add null check for EmailHeaderSame
When asserting that an email header matches some text and the header
does not exist on the message, an `Error` was being thrown because the
method `getBodyAsString()` was being called on a `null` object.  This
commit allows a missing header and will show the user that the header's
value is `null`.

Issue #44190