[Bridge] Corrects bug in test listener trait
If you enable the `SymfonyTestsListenerTrait`, use the
`ExpectDeprecationTrait`, and call
`$this->expectNotToPerformAssertions()`, an error was returned in your
test saying it was risky as it did not perform any assertions even
though it was expected not to.

This bug turned out to be caused by not checking if the
`doesNotPerformAssertions` flag had been set.

Issue: #41444